+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2.1.5
 - Fix and Add Missing Mossy Red Rock and Mossy Stone Recipes
 - Add Missing Foragers Table Recipe
+- Use Common Dyes Tags for Colored Sand Recipes
 
 # 2.1.4 (Forge Only)
 - Require Forge 52.0.20 as a minimum due to the addition of new Forge tags and implementation of Common Tags