+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - Make Entity Model Classes Package-Private, remove pointless override of getAnimation Method
 - Move PoseStack Scaling into preRender Method
 - Decrease Pumpkin Warden Shadow Radius
+- Fix Incompatibility with William Wythers' Overhauled Overworld/Expanded Ecosphere
+  - Remove SeaGrass Normal from Lush Stacks as SeaGrass Warm is the same feature but more common 
 
 # 2.1.4 (Forge Only)
 - Require Forge 52.0.20 as a minimum due to the addition of new Forge tags and implementation of Common Tags